Fixed a MAJOR bug where the number of particles were inproperly allocated when initially read in.

+++ include/halo_extraction_classes.h
@@ -347,6 +347,6 @@ particlePosition::particlePosition(){
 }
 
 
-extern "C" unsigned long readpmss_( int *jstep, char *filestart, int *filestartlength );
+extern "C" long long readpmss_( int *jstep, char *filestart, int *filestartlength );
 
 #endif // HE_CLASSES
